[refactor] : (코드 변경 x) atoms & templates 컴포넌트 명칭 및 기능별 카테고리화
