--- src/components/organisms/Info.jsx
@@ -1,45 +1,0 @@
-import Time from '../atoms/Time';
-
-/* eslint-disable */
-const Info = ({ response }) => {
-  // 마감시간
-  let finishTime = new Date(response.finishedAt);
-
-  // 메뉴들 나열하는 거
-  const showMenu = (item) => {
-    return item.map(({ name }) => {
-      return <div key={name}>{name}</div>;
-    });
-  };
-
-  return (
-    <div className="px-8 pt-6">
-      {/* 주문정보 */}
-      <div>
-        <div className="text-xl font-bold text-blue py-2">주문 정보</div>
-        <div>{showMenu(response.beverage)}</div>
-      </div>
-      {/* 요청사항 */}
-      <div className="my-12">
-        <div className="text-xl font-bold text-blue py-2">요청 사항</div>
-        <div>
-          <div className="flex">
-            <div className="text-zinc-400">픽업팁</div>
-            <div className="ml-8">{response.tip}원</div>
-          </div>
-          <div className="flex">
-            <div className="text-zinc-400">요청사항</div>
-            <div className="ml-5">{response.request}</div>
-          </div>
-        </div>
-      </div>
-      {/* 마감기한 */}
-      <div className="my-12">
-        <div className="text-xl font-bold text-blue py-2">마감기한</div>
-        <div>{<Time date={finishTime} />}</div>
-      </div>
-    </div>
-  );
-};
-
-export default Info;
