+++ src/components/templates/mypage/AdminMyPageTemplate.jsx
@@ -2,9 +2,9 @@ import { CiEdit } from 'react-icons/ci';
 import { PiCaretRightLight, PiReadCvLogoThin, PiCheckSquareLight } from 'react-icons/pi';
 import { RiAdminLine } from 'react-icons/ri';
 import { useNavigate } from 'react-router-dom';
-import routes from '../../constant/routes';
+import routes from '../../../constant/routes';
 
-const AdminMyPage = () => {
+const AdminMyPageTemplate = () => {
   const navigate = useNavigate();
   const navigateAdmin = () => {
     return navigate(routes.admin);
@@ -42,4 +42,4 @@ const AdminMyPage = () => {
   );
 };
 
-export default AdminMyPage;
+export default AdminMyPageTemplate;
