+++ src/components/templates/mypage/GuestMyPageTemplate.jsx
@@ -1,9 +1,9 @@
 import { CiEdit } from 'react-icons/ci';
 import { PiGraduationCapLight, PiCaretRightLight } from 'react-icons/pi';
 import { useNavigate } from 'react-router-dom';
-import routes from '../../constant/routes';
+import routes from '../../../constant/routes';
 
-const GuestMyPage = () => {
+const GuestMyPageTemplate = () => {
   const navigate = useNavigate();
   const uploadStudentIdCard = () => {
     return navigate(routes.uploadStudentCard);
@@ -31,4 +31,4 @@ const GuestMyPage = () => {
   );
 };
 
-export default GuestMyPage;
+export default GuestMyPageTemplate;
