+++ src/pages/PostWritePage.jsx
@@ -5,9 +5,9 @@ import { useMutation } from '@tanstack/react-query';
 import Swal from 'sweetalert2';
 import OtherNav from '../components/atoms/nav/OtherNav';
 import BtnNavigate from '../components/molecules/BtnNavigate';
-import OrderInfo from '../components/templates/OrderInfo';
-import OrderRequest from '../components/templates/OrderRequest';
-import OrderDeadLine from '../components/templates/OrderDeadLine';
+import OrderInfo from '../components/templates/postWrite/OrderInfoTemplate';
+import OrderRequest from '../components/templates/postWrite/OrderRequestTemplate';
+import OrderDeadLine from '../components/templates/postWrite/OrderDeadLineTemplate';
 import CircleNavigate from '../components/organisms/CircleNavigate';
 import { STORE, BEVERAGE } from '../constant/postWrite/orderInfo';
 import { DESTINATION } from '../constant/postWrite/orderRequest';
